docs: refactor GetUserIDFromRequest() and JwtSecret to global
