+++ backend/tokens/jwt.go
@@ -1,0 +1,47 @@
+package tokens
+
+import (
+	"fmt"
+	"log"
+	"net/http"
+	"os"
+
+	"github.com/golang-jwt/jwt"
+)
+
+var JwtSecret = []byte(os.Getenv("JWT_SECRET"))
+
+func init() {
+	s := os.Getenv("JWT_SECRET")
+	if s == "" {
+		log.Fatal("JWT_SECRET is not set")
+	}
+	JwtSecret = []byte(s)
+}
+
+func GetUserIDFromRequest(r *http.Request) (int, error) {
+	authHeader := r.Header.Get("Authorization")
+	if authHeader == "" {
+		return 0, fmt.Errorf("authorization header missing")
+	}
+	var tokenString string
+	fmt.Sscanf(authHeader, "Bearer %s", &tokenString)
+	if tokenString == "" {
+		return 0, fmt.Errorf("token missing in authorization header")
+	}
+	token, err := jwt.Parse(tokenString, func(token *jwt.Token) (interface{}, error) {
+		return JwtSecret, nil
+	})
+	if err != nil || !token.Valid {
+		return 0, fmt.Errorf("invalid token: %v", err)
+	}
+	claims, ok := token.Claims.(jwt.MapClaims)
+	if !ok {
+		return 0, fmt.Errorf("invalid token claims")
+	}
+	userIDFloat, ok := claims["user_id"].(float64)
+	if !ok {
+		return 0, fmt.Errorf("user_id not found in token claims")
+	}
+	return int(userIDFloat), nil
+}
